backend fixes

+++ Backend/library-web-backend/src/main/java/com/example/librarywebbackend/service/implementation/AuthorService.java
@@ -2,6 +2,7 @@ package com.example.librarywebbackend.service.implementation;
 
 import com.example.librarywebbackend.entity.Author;
 import com.example.librarywebbackend.repository.AuthorRepository;
+import com.example.librarywebbackend.repository.BookAuthorRepository;
 import com.example.librarywebbackend.service.IAuthorService;
 import org.springframework.stereotype.Service;
 
@@ -12,8 +13,12 @@ public class AuthorService implements IAuthorService {
 
     private final AuthorRepository authorRepository;
 
-    public AuthorService(AuthorRepository authorRepository) {
+    private final BookAuthorRepository bookAuthorRepository;
+
+
+    public AuthorService(AuthorRepository authorRepository, BookAuthorRepository bookAuthorRepository) {
         this.authorRepository = authorRepository;
+        this.bookAuthorRepository = bookAuthorRepository;
     }
 
     @Override
@@ -46,8 +51,20 @@ public class AuthorService implements IAuthorService {
 
     @Override
     public void deleteAuthor(Long id) {
+
+        Author author = authorRepository.findById(id)
+                .orElseThrow(() -> new RuntimeException("Author not found"));
+
+        // sprawdzamy powiązania
+        int count = bookAuthorRepository.countByAuthorId(id);
+
+        if (count > 0) {
+            throw new IllegalStateException("Cannot delete author. Author is assigned to " + count + " books.");
+        }
+
         authorRepository.deleteById(id);
     }
+
 
     @Override
     public List<Author> searchAuthors(String query) {
